@@ -1,13 +1,14 @@
-using Newtonsoft.Json;
+
+using System.Text.Json.Serialization;
 
 namespace RetireNet.Runtimes.Core.Clients.Models
 {
     public class CVE
     {
-        [JsonProperty("cve-id")]
+        [JsonPropertyName("cve-id")]
         public string Id { get; set; }
 
-        [JsonProperty("cve-url")]
+        [JsonPropertyName("cve-url")]
         public string Url { get; set; }
     }
 }
