@@ -1,11 +1,11 @@
 using System;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 namespace RetireNet.Runtimes.Core.HttpClients.Models.Index
 {
     internal class Channel
     {
-        [JsonProperty("releases.json")]
+        [JsonPropertyName("releases.json")]
         public Uri ReleasesUrl { get; set; }
     }
 }
