@@ -56,7 +56,7 @@ namespace RetireNet.Runtimes.Core.Clients.Models
         private static string GetRuntimeVersion(string framework)
         {
             var assembly = typeof(System.Runtime.GCSettings).GetTypeInfo().Assembly;
-            var assemblyPath = assembly.CodeBase.Split(new[] {'/', '\\'}, StringSplitOptions.RemoveEmptyEntries);
+            var assemblyPath = assembly.Location.Split(new[] {'/', '\\'}, StringSplitOptions.RemoveEmptyEntries);
             int netCoreAppIndex = Array.IndexOf(assemblyPath, framework);
             if (netCoreAppIndex > 0 && netCoreAppIndex < assemblyPath.Length - 2)
                 return assemblyPath[netCoreAppIndex + 1];
